backend/requirements.txt
alembic==1.14.0
bleach==6.2.0
blinker==1.9.0
cache==1.0.3
cachelib==0.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
cors==1.0.1
filelock==3.16.1
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
future==1.0.0
gevent==24.11.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.6
MarkupSafe==3.0.2
marshmallow==3.23.1
migrate==0.3.8
packaging==24.2
passlib==1.7.4
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.10.0
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
requests-file==2.1.0
setuptools==75.6.0
SQLAlchemy==2.0.36
tldextract==5.1.3
typing_extensions==4.12.2
urllib3==2.2.3
webencodings==0.5.1
Werkzeug==3.1.3
zope.event==5.0
zope.interface==7.2
